+++ lukehog/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.1
+
+- add Lukehog.useAnonymousUserId()
+- fix parsing error
+- remove auto retry
+
 ## 0.2.0
 
 - Make appId constructor argument positional